Align op namespace with DataAPI (Matmul, Cast -> Astype)
Separating Matmul into "linear algebra" namespace (with "Var" reverted)

Astype
