+++ python/tvm/relax/op/__init__.py
@@ -21,6 +21,7 @@
 from .base import *
 from .create import *
 from .datatype import *
+from .linear_algebra import *
 from .manipulate import *
 from .op_attrs import *
 from .statistical import *
