+++ python/tvm/script/ir_builder/relax/ir.py
@@ -30,9 +30,9 @@ from tvm.relax.analysis import get_static_type
 from tvm.relax.op import (
     add,
     assert_op,
+    astype,
     builtin,
     call_tir,
-    cast,
     concat,
     cos,
     divide,
@@ -47,6 +47,7 @@ from tvm.relax.op import (
     less,
     log,
     make_closure,
+    matmul,
     max,
     mean,
     memory,
@@ -432,10 +433,10 @@ __all__ = [
     "add",
     "arg",
     "assert_op",
+    "astype",
     "builtin",
     "call_packed",
     "call_tir",
-    "cast",
     "concat",
     "const",
     "cos",
@@ -459,6 +460,7 @@ __all__ = [
     "less",
     "log",
     "make_closure",
+    "matmul",
     "max",
     "mean",
     "memory",
