+++ src/relax/op/nn/nn.cc
@@ -220,96 +220,6 @@ TVM_REGISTER_OP("relax.nn.layer_norm")
     .add_argument("beta", "Tensor", "The beta offset factor.")
     .set_attr<FInferStructInfo>("FInferStructInfo", InferStructInfoLayerNorm);
 
-/* relax.nn.matmul */
-TVM_REGISTER_NODE_TYPE(MatmulAttrs);
-
-Expr Matmul(Expr a, Expr b, DataType out_dtype) {
-  ObjectPtr<MatmulAttrs> attrs = make_object<MatmulAttrs>();
-  attrs->out_dtype = out_dtype;
-
-  static const Op& op = Op::Get("relax.nn.matmul");
-  return Call(op, {std::move(a), std::move(b)}, Attrs(attrs), {});
-}
-
-TVM_REGISTER_GLOBAL("relax.op.nn.matmul").set_body_typed(Matmul);
-
-StructInfo InferStructInfoMatmul(const Call& call, const BlockBuilder& ctx) {
-  Array<TensorStructInfo> input_sinfo = GetInputTensorStructInfo(call, ctx);
-  TensorStructInfo lhs_sinfo = input_sinfo[0];
-  TensorStructInfo rhs_sinfo = input_sinfo[1];
-
-  const auto* attrs = call->attrs.as<MatmulAttrs>();
-  DataType out_dtype = attrs->out_dtype.is_void()
-                           ? InferBinaryArithOpOutDtype(call, ctx, lhs_sinfo, rhs_sinfo)
-                           : attrs->out_dtype;
-
-  if (lhs_sinfo->IsUnknownNdim() || rhs_sinfo->IsUnknownNdim()) {
-    return TensorStructInfo(out_dtype, kUnknownNDim);
-  }
-  int lhs_ndim = lhs_sinfo->ndim;
-  int rhs_ndim = rhs_sinfo->ndim;
-  if (lhs_ndim == 0 || rhs_ndim == 0) {
-    ctx->ReportFatal(Diagnostic::Error(call)
-                     << "Matmul requires both inputs to have at least 1 dimension. However, "
-                     << (lhs_ndim == 0 ? "lhs" : "rhs") << " is a 0-rank tensor.");
-  }
-
-  int lhs_prepended = 0;
-  int rhs_appended = 0;
-  if (lhs_ndim == 1) {
-    lhs_ndim = 2;
-    lhs_prepended = 1;
-  }
-  if (rhs_ndim == 1) {
-    rhs_ndim = 2;
-    rhs_appended = 1;
-  }
-  int output_ndim = std::max(lhs_ndim, rhs_ndim) - lhs_prepended - rhs_appended;
-
-  const auto* lhs_shape = lhs_sinfo->shape.as<ShapeExprNode>();
-  const auto* rhs_shape = rhs_sinfo->shape.as<ShapeExprNode>();
-  if (lhs_shape == nullptr || rhs_shape == nullptr) {
-    return TensorStructInfo(out_dtype, output_ndim);
-  }
-
-  Array<PrimExpr> lhs_shape_prefix{lhs_shape->values.begin(),
-                                   lhs_shape->values.end() - 2 + lhs_prepended};
-  Array<PrimExpr> rhs_shape_prefix{rhs_shape->values.begin(),
-                                   rhs_shape->values.end() - 2 + rhs_appended};
-  Optional<Array<PrimExpr>> output_shape_prefix =
-      InferBinaryBroadcastShape(call, ctx, lhs_shape_prefix, rhs_shape_prefix);
-  if (!output_shape_prefix.defined()) {
-    return TensorStructInfo(out_dtype, output_ndim);
-  }
-
-  arith::Analyzer* analyzer = ctx->GetAnalyzer();
-  PrimExpr lhs_reduction_length = lhs_shape->values[lhs_sinfo->ndim - 1];
-  PrimExpr rhs_reduction_length = rhs_shape->values[rhs_ndim - 2];
-  if (analyzer->CanProve(lhs_reduction_length != rhs_reduction_length)) {
-    ctx->ReportFatal(Diagnostic::Error(call)
-                     << "Matmul requires the reduction length of lhs and rhs to be equal. However, "
-                        "the reduction lengths of lhs and rhs are "
-                     << lhs_reduction_length << " and " << rhs_reduction_length
-                     << " respectively.");
-  }
-
-  Array<PrimExpr> output_shape = output_shape_prefix.value();
-  if (!lhs_prepended) {
-    output_shape.push_back(lhs_shape->values[lhs_ndim - 2]);
-  }
-  if (!rhs_appended) {
-    output_shape.push_back(rhs_shape->values[rhs_ndim - 1]);
-  }
-  ICHECK_EQ(static_cast<int>(output_shape.size()), output_ndim);
-  return TensorStructInfo(ShapeExpr(output_shape), out_dtype);
-}
-
-TVM_REGISTER_OP("relax.nn.matmul")
-    .set_num_inputs(2)
-    .add_argument("a", "Tensor", "The left operand of the matmul.")
-    .add_argument("b", "Tensor", "The right operand of the matmul.")
-    .set_attr<FInferStructInfo>("FInferStructInfo", InferStructInfoMatmul);
-
 /* relax.nn.dropout */
 TVM_REGISTER_NODE_TYPE(DropoutAttrs);
