+++ tests/python/relax/test_op_datatype.py
@@ -28,7 +28,7 @@ import numpy as np
 def test_op_correctness():
     x = relax.Var("x", R.Tensor((2, 3), "float32"))
     c = relax.Constant(tvm.nd.array(np.array([1, 2, 3], dtype="float16")))
-    assert relax.op.cast(x, "float16").op == Op.get("relax.cast")
+    assert relax.op.astype(x, "float16").op == Op.get("relax.astype")
     assert relax.op.wrap_param(c, "float32").op == Op.get("relax.wrap_param")
 
 
@@ -37,7 +37,7 @@ def _check_inference(bb: relax.BlockBuilder, call: relax.Call, expected_sinfo: r
     tvm.ir.assert_structural_equal(ret.struct_info, expected_sinfo)
 
 
-def test_cast_infer_struct_info():
+def test_astype_infer_struct_info():
     bb = relax.BlockBuilder()
     x0 = relax.Var("x", R.Tensor((2, 3), "float32"))
     x1 = relax.Var("x", R.Tensor("float32", ndim=2))
@@ -46,30 +46,30 @@ def test_cast_infer_struct_info():
     x4 = relax.Var("x", R.Tensor(ndim=2))
     x5 = relax.Var("x", R.Tensor())
 
-    _check_inference(bb, relax.op.cast(x0, "float16"), relax.TensorStructInfo((2, 3), "float16"))
+    _check_inference(bb, relax.op.astype(x0, "float16"), relax.TensorStructInfo((2, 3), "float16"))
     _check_inference(
-        bb, relax.op.cast(x1, "float16"), relax.TensorStructInfo(dtype="float16", ndim=2)
+        bb, relax.op.astype(x1, "float16"), relax.TensorStructInfo(dtype="float16", ndim=2)
     )
-    _check_inference(bb, relax.op.cast(x2, "float16"), relax.TensorStructInfo(dtype="float16"))
-    _check_inference(bb, relax.op.cast(x3, "float16"), relax.TensorStructInfo((2, 3), "float16"))
+    _check_inference(bb, relax.op.astype(x2, "float16"), relax.TensorStructInfo(dtype="float16"))
+    _check_inference(bb, relax.op.astype(x3, "float16"), relax.TensorStructInfo((2, 3), "float16"))
     _check_inference(
-        bb, relax.op.cast(x4, "float16"), relax.TensorStructInfo(dtype="float16", ndim=2)
+        bb, relax.op.astype(x4, "float16"), relax.TensorStructInfo(dtype="float16", ndim=2)
     )
-    _check_inference(bb, relax.op.cast(x5, "float16"), relax.TensorStructInfo(dtype="float16"))
+    _check_inference(bb, relax.op.astype(x5, "float16"), relax.TensorStructInfo(dtype="float16"))
 
 
-def test_cast_infer_struct_info_shape_symbolic():
+def test_astype_infer_struct_info_shape_symbolic():
     bb = relax.BlockBuilder()
     m = tir.Var("m", "int64")
     n = tir.Var("n", "int64")
     x0 = relax.Var("x", R.Tensor((m, n), "float32"))
     x1 = relax.Var("x", R.Tensor((m, n)))
 
-    _check_inference(bb, relax.op.cast(x0, "float16"), relax.TensorStructInfo((m, n), "float16"))
-    _check_inference(bb, relax.op.cast(x1, "float16"), relax.TensorStructInfo((m, n), "float16"))
+    _check_inference(bb, relax.op.astype(x0, "float16"), relax.TensorStructInfo((m, n), "float16"))
+    _check_inference(bb, relax.op.astype(x1, "float16"), relax.TensorStructInfo((m, n), "float16"))
 
 
-def test_cast_infer_struct_info_shape_var():
+def test_astype_infer_struct_info_shape_var():
     bb = relax.BlockBuilder()
     s0 = relax.Var("s", relax.ShapeStructInfo((2, 3)))
     s1 = relax.Var("s", relax.ShapeStructInfo(ndim=2))
@@ -78,39 +78,41 @@ def test_cast_infer_struct_info_shape_var():
     x1 = relax.Var("x", relax.TensorStructInfo(s1, "float32"))
     x2 = relax.Var("x", relax.TensorStructInfo(s2, "float32"))
 
-    _check_inference(bb, relax.op.cast(x0, "float16"), relax.TensorStructInfo(s0, "float16"))
-    _check_inference(bb, relax.op.cast(x1, "float16"), relax.TensorStructInfo(s1, "float16"))
-    _check_inference(bb, relax.op.cast(x2, "float16"), relax.TensorStructInfo(s2, "float16"))
+    _check_inference(bb, relax.op.astype(x0, "float16"), relax.TensorStructInfo(s0, "float16"))
+    _check_inference(bb, relax.op.astype(x1, "float16"), relax.TensorStructInfo(s1, "float16"))
+    _check_inference(bb, relax.op.astype(x2, "float16"), relax.TensorStructInfo(s2, "float16"))
 
 
-def test_cast_infer_struct_info_more_input_dtype():
+def test_astype_infer_struct_info_more_input_dtype():
     bb = relax.BlockBuilder()
     x0 = relax.Var("x", R.Tensor((2, 3), "float16"))
     x1 = relax.Var("x", R.Tensor((2, 3), "int8"))
     x2 = relax.Var("x", R.Tensor((2, 3), "int32"))
 
-    _check_inference(bb, relax.op.cast(x0, "float32"), relax.TensorStructInfo((2, 3), "float32"))
-    _check_inference(bb, relax.op.cast(x1, "int32"), relax.TensorStructInfo((2, 3), "int32"))
-    _check_inference(bb, relax.op.cast(x2, "int8"), relax.TensorStructInfo((2, 3), "int8"))
+    _check_inference(bb, relax.op.astype(x0, "float32"), relax.TensorStructInfo((2, 3), "float32"))
+    _check_inference(bb, relax.op.astype(x1, "int32"), relax.TensorStructInfo((2, 3), "int32"))
+    _check_inference(bb, relax.op.astype(x2, "int8"), relax.TensorStructInfo((2, 3), "int8"))
 
 
-def test_cast_infer_struct_info_wrong_input_type():
+def test_astype_infer_struct_info_wrong_input_type():
     bb = relax.BlockBuilder()
     x0 = relax.Var("x", relax.ShapeStructInfo((2, 3)))
     x1 = relax.Var("x", relax.FuncStructInfo([], R.Tensor((2, 3), "float32")))
 
     with pytest.raises(TVMError):
-        bb.normalize(relax.op.cast(x0, "float16"))
+        bb.normalize(relax.op.astype(x0, "float16"))
     with pytest.raises(TVMError):
-        bb.normalize(relax.op.cast(x1, "float16"))
+        bb.normalize(relax.op.astype(x1, "float16"))
 
 
 def test_wrap_param_infer_struct_info():
     bb = relax.BlockBuilder()
     x0 = relax.Constant(tvm.nd.array(np.zeros([1, 2, 3], dtype="float16")))
     x1 = relax.Constant(tvm.nd.array(np.zeros([1, 2, 3], dtype="int8")))
-    _check_inference(bb, relax.op.cast(x0, "float32"), relax.TensorStructInfo((1, 2, 3), "float32"))
-    _check_inference(bb, relax.op.cast(x1, "int32"), relax.TensorStructInfo((1, 2, 3), "int32"))
+    _check_inference(
+        bb, relax.op.astype(x0, "float32"), relax.TensorStructInfo((1, 2, 3), "float32")
+    )
+    _check_inference(bb, relax.op.astype(x1, "int32"), relax.TensorStructInfo((1, 2, 3), "int32"))
 
 
 if __name__ == "__main__":
